+++ NEWS.md
@@ -1,5 +1,8 @@
 # tern.mmrm 0.1.1.9003
 
+* Added option `average_emmeans` to `fit_mmrm()` which allows to estimate
+  least square means for averages of visits.
+
 # tern.mmrm 0.1.1
 
 * Removed dependency on `purrr`, `tibble`, `scda` and `scda.2022`